refactor: move functions out of lib
